+++ src/components/Badge/Badge.helpers.tsx
@@ -9,10 +9,7 @@ const variants: Record<BadgeVariant, GetVariantStylesFn> = {
       {
         background: theme.tones[tone].superLight,
         border: `1px solid ${theme.tones[tone].lighter}`,
-        color: theme.tones[tone].darker,
-      },
-      (tone === "WARNING" || tone === "SUCCESS") && {
-        color: theme.tones[tone].superDark,
+        color: theme.tones[tone].text,
       },
     ]
   },
@@ -20,14 +17,9 @@ const variants: Record<BadgeVariant, GetVariantStylesFn> = {
     return theme => [
       {
         background: theme.tones[tone].dark,
-        color: theme.colors.white,
-      },
-      tone === `WARNING` && {
-        background: theme.colors.yellow[50],
-        color: theme.colors.blackFade[70],
-      },
-      tone === `SUCCESS` && {
-        background: theme.colors.green[80],
+        color: theme.tones[tone].textInverted
+          ? theme.tones[tone].textInverted
+          : theme.colors.white,
       },
     ]
   },