+++ src/components/Notification/Notification.tsx
@@ -26,6 +26,7 @@ const NotificationContext = React.createContext<NotificationContextValue>({
 })
 
 const baseCss: ThemeCss = theme => ({
+  color: theme.colors.grey[90],
   display: `flex`,
   alignItems: `flex-start`,
   justifyContent: `space-between`,
@@ -88,12 +89,12 @@ export default function Notification({
         {content && (
           <NotificationContent
             as={contentAs}
-            css={theme => ({
-              color:
-                variant === `SOLID` && tone !== `WARNING`
-                  ? theme.colors.white
-                  : null,
-            })}
+            css={theme => [
+              linkUrl &&
+                linkText && {
+                  marginRight: theme.space[5],
+                },
+            ]}
           >
             {Icon && (
               <Icon
@@ -107,12 +108,10 @@ export default function Notification({
                     height: "1em",
                   },
                   variant === `SOLID` && {
-                    color: theme.colors.whiteFade[90],
+                    color: theme.tones[tone].mediumInverted
+                      ? theme.tones[tone].mediumInverted
+                      : theme.colors.whiteFade[90],
                   },
-                  variant === `SOLID` &&
-                    tone === `WARNING` && {
-                      color: theme.colors.blackFade[90],
-                    },
                 ]}
               />
             )}
@@ -125,6 +124,9 @@ export default function Notification({
             to={linkUrl}
             onClick={onLinkClick}
             css={theme => ({
+              // to push <Link> to the right also when there's a
+              // <NotificationDismissButton>
+              marginLeft: "auto",
               color: variant === `SOLID` ? theme.colors.white : null,
               ":hover": {
                 color: variant === `SOLID` ? theme.colors.whiteFade[80] : null,
@@ -163,11 +165,10 @@ function NotificationContent({
 }: NotificationContentProps) {
   return (
     <Component
-      css={(theme: Theme) => ({
+      css={{
         display: `flex`,
         alignItems: `flex-start`,
-        color: theme.tones[`NEUTRAL`].superDark,
-      })}
+      }}
       {...rest}
     />
   )